Day2-hands-on/hf_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
python-jose[cryptography]
python-dotenv
requests
huggingface_hub
